Chore: moved tests into correct dir, and fix references
