+++ src/components/__tests__/MapPanel.spec.jsx
@@ -1,8 +1,8 @@
 import configureMockStore from 'redux-mock-store'
 import { IntlProvider } from 'react-intl'
 import { Provider } from 'react-redux'
-import { MapPanel, mapDispatchToProps } from '../components/Map/MapPanel'
-import { MODE_MAP } from '../constants'
+import { MapPanel, mapDispatchToProps } from '../Map/MapPanel'
+import { MODE_MAP } from '../../constants'
 import React from 'react'
 import renderer from 'react-test-renderer'
 import thunk from 'redux-thunk'
