+++ src/components/__tests__/Pill.spec.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import { shallow } from 'enzyme';
-import { SLUG_SEPARATOR } from '../constants'
-import { Pill, mapStateToProps, mapDispatchToProps } from '../components/Search/Pill';
+import { SLUG_SEPARATOR } from '../../constants'
+import { Pill, mapStateToProps, mapDispatchToProps } from '../Search/Pill';
 
 function setupEnzyme() {
   const props = {
