+++ src/components/__tests__/PrintInfo.spec.jsx
@@ -1,6 +1,6 @@
 import {
   getComplaintCountText, mapStateToProps, PrintInfo
-} from '../components/Print/PrintInfo'
+} from '../Print/PrintInfo'
 import { IntlProvider } from 'react-intl'
 import React from 'react'
 import renderer from 'react-test-renderer'
