+++ src/components/__tests__/SearchComponents.spec.jsx
@@ -1,4 +1,4 @@
-import { mapStateToProps } from '../components/Search/SearchComponents'
+import { mapStateToProps } from '../Search/SearchComponents'
 import React from 'react'
 
 describe( 'component: SearchComponents', () => {
