+++ src/components/__tests__/SearchPanel.spec.jsx
@@ -1,4 +1,4 @@
-import { SearchPanel } from '../components/Search/SearchPanel'
+import { SearchPanel } from '../Search/SearchPanel'
 import React from 'react'
 import renderer from 'react-test-renderer'
 import thunk from 'redux-thunk'
