+++ src/components/__tests__/UrlBarSynch.spec.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { UrlBarSynch, mapDispatchToProps } from '../components/UrlBarSynch';
-import * as types from '../constants'
+import { UrlBarSynch, mapDispatchToProps } from '../UrlBarSynch';
+import * as types from '../../constants'
 
 describe('component:UrlBarSynch', () =>{
   let target;
